Implement affine gap penalties.
This required changing most values to doubles instead of integers;
the header values for the N and W matrices can then be expressed
as Double.NEGATIVE_INFINITY. Using an integer in this same situation
caused overflow and incorrect alignment results.

+++ src/alignment/AlignmentElement.java
@@ -5,15 +5,15 @@ package alignment;
 
 public class AlignmentElement
 {
-	public Integer score;
+	public Double score;
 	public PointerDirection direction;
 	
-	public AlignmentElement(Integer score_)
+	public AlignmentElement(Double score_)
 	{
 		score = score_;
 	}
 	
-	public AlignmentElement(Integer score_, PointerDirection direction_)
+	public AlignmentElement(Double score_, PointerDirection direction_)
 	{
 		score = score_;
 		direction = direction_;
